requirement.txt
lmstudio
requests
uuid
datetime
websocket-client
openai
rp
